task2_asr_transliteration/requirements.txt
# ASR
openai-whisper==20250625

# Transliteration
indic-transliteration>=2.3.57

# Interface
gradio>=4.44.0

# Audio processing
ffmpeg-python==0.2.0

# ML / Torch
torch>=2.1.0
torchaudio>=2.1.0

# Utilities
numpy>=1.26.4

# Testing
pytest>=9.0.0